add address hash function; refactor iproperty file; add test
