+++ python/shared/iproperty.py
@@ -1,9 +1,10 @@
 from enum import Enum
 from datetime import datetime, timezone
 
-import usaddress # type: ignore
 import uuid
 import logging
+
+from shared.iproperty_address import IPropertyAddress
 
 class AreaUnit(Enum):
     SquareFeet = "SquareFeet"
@@ -23,20 +24,13 @@ class PropertyArea:
     def __str__(self):
         return f"{self.area} {self.unit.value}"
 
-class AddressType(Enum):
-    StreetAddress = "Street Address"
-    Intersection = "Intersection"
-    POBox = "PO Box"
-    Ambiguous = "Ambiguous"
-
 class PropertyType(Enum):
     SingleFamily = "Single-family"
     Townhome = "Townhome"
     Condo = "Condo"
 
-# Concat different tags into a single street address
-# usaddress doc: https://parserator.datamade.us/api-docs/
-def extractStreetAddress(addressPropertyBag: dict) -> str:
+# Deprecated
+def _extractStreetAddress(addressPropertyBag: dict) -> str:
     addressTags = set({
         "AddressNumber",
         "AddressNumberPrefix",
@@ -61,162 +55,6 @@ def extractStreetAddress(addressPropertyBag: dict) -> str:
         if key in addressTags:
             streetAddress += (" " if len(streetAddress) > 0 else "") + value
     return streetAddress
-
-def extractUnitInformation(addressPropertyBag: dict) -> str:
-    addressTags = set({
-        "OccupancyIdentifier",
-        "OccupancyType",
-    })
-    unit: str = ""
-    for key, value in addressPropertyBag.items():
-        if key in addressTags:
-            unit += (" " if len(unit) > 0 else "") + value
-    return unit
-
-# Convert address string to a hash string
-def get_address_hash(address: str, logger: logging.Logger | None = None) -> str:
-    # USPS standard abbreviations for street suffixes and directionals
-    suffix_abbr = {
-        "street": "St",
-        "st": "St",
-        "avenue": "Ave",
-        "ave": "Ave",
-        "boulevard": "Blvd",
-        "blvd": "Blvd",
-        "road": "Rd",
-        "rd": "Rd",
-        "drive": "Dr",
-        "dr": "Dr",
-        "lane": "Ln",
-        "ln": "Ln",
-        "court": "Ct",
-        "ct": "Ct",
-        "place": "Pl",
-        "pl": "Pl",
-        "terrace": "Ter",
-        "ter": "Ter",
-        "circle": "Cir",
-        "cir": "Cir",
-        "parkway": "Pkwy",
-        "pkwy": "Pkwy",
-        "way": "Way",
-        "trail": "Trl",
-        "trl": "Trl",
-        "highway": "Hwy",
-        "hwy": "Hwy",
-        "driveway": "Dr",
-        "drwy": "Dr",
-        # Add more as needed
-    }
-    directional_abbr = {
-        "north": "N",
-        "n": "N",
-        "south": "S",
-        "s": "S",
-        "east": "E",
-        "e": "E",
-        "west": "W",
-        "w": "W",
-        "northeast": "NE",
-        "ne": "NE",
-        "northwest": "NW",
-        "nw": "NW",
-        "southeast": "SE",
-        "se": "SE",
-        "southwest": "SW",
-        "sw": "SW",
-    }
-    def abbr_word(word: str, abbr_map: dict) -> str:
-        return abbr_map.get(word.lower(), word)
-    try:
-        parsedAddress = usaddress.tag(address)
-        addressPropertyBag = parsedAddress[0]
-        # Build street part with abbreviation normalization
-        street_components = [
-            addressPropertyBag.get("AddressNumber", ""),
-            addressPropertyBag.get("AddressNumberPrefix", ""),
-            abbr_word(addressPropertyBag.get("StreetNamePreDirectional", ""), directional_abbr),
-            addressPropertyBag.get("StreetNamePreType", ""),
-            addressPropertyBag.get("StreetName", ""),
-            abbr_word(addressPropertyBag.get("StreetNamePostType", ""), suffix_abbr),
-            abbr_word(addressPropertyBag.get("StreetNamePostDirectional", ""), directional_abbr)
-        ]
-        street = " ".join(filter(None, street_components))
-        # Build unit part
-        unit_components = [
-            addressPropertyBag.get("OccupancyType", ""),
-            addressPropertyBag.get("OccupancyIdentifier", "")
-        ]
-        unit = " ".join(filter(None, unit_components))
-        # City, state, zip
-        city = addressPropertyBag.get("PlaceName", "")
-        state = addressPropertyBag.get("StateName", "")
-        zipcode = addressPropertyBag.get("ZipCode", "")
-        # Concatenate all parts
-        components = [street]
-        if unit:
-            components.append(unit)
-        components.extend([city, state, zipcode])
-        normalized = ",".join(filter(None, components))
-        # Apply normalization: lowercase, spaces to '-', commas to '|'
-        normalized = normalized.lower().replace(" ", "-").replace(",", "|")
-        return normalized
-    except Exception as e:
-        # Fallback: normalize the original string if parsing fails
-        error_msg = f"Error parsing address: {address}, error: {e}"
-        if logger != None:
-            logger.error(error_msg)
-        else:
-            print(error_msg)
-        raise Exception(error_msg)
-
-# TODO: Use USPS address format API?
-class IPropertyAddress:
-    def __init__(self, address: str):
-        parsedAddress = usaddress.tag(address)
-        addressType: str = parsedAddress[1]
-        if (addressType != AddressType.StreetAddress.value):
-            raise ValueError(f"Invalid address type: {addressType} for address: {address}")
-        addressPropertyBag: dict = parsedAddress[0]
-        self._streetName: str = extractStreetAddress(addressPropertyBag)
-        self._unit: str = extractUnitInformation(addressPropertyBag)
-        self._state: str = addressPropertyBag["StateName"]
-        self._city: str = addressPropertyBag["PlaceName"]
-        self._zipCode: str = addressPropertyBag["ZipCode"]
-
-        self._fullAddress: str = self._streetName + "," + ((self._unit + ",") if len(self._unit) > 0 else "") + self._city + "," + self._state + "," + self._zipCode
-
-    @property
-    def streetName(self) -> str:
-        return self._streetName
-
-    @property
-    def unit(self) -> str:
-        return self._unit
-
-    @property
-    def state(self) -> str:
-        return self._state
-
-    @property
-    def city(self) -> str:
-        return self._city
-
-    @property
-    def zipCode(self) -> str:
-        return self._zipCode
-
-    def getAddressLine(self) -> str:
-        return self._fullAddress
-
-    # This is index related
-    def __eq__(self, other):
-        if not isinstance(other, IPropertyAddress):
-            return NotImplemented
-        return self._fullAddress == other._fullAddress
-
-    def __str__(self):
-        return f"Full address: {self._fullAddress}, Street: {self.streetName}, UnitNumber(if any): {self._unit}, State: {self._state}, ZipCode: {self._zipCode}"
 
 class IPropertyHistoryEventType(Enum):
     Listed = "Listed"
